bot/requirements.txt
aiogram==3.22.0
asyncpg==0.30.0
python-dotenv==1.1.1
APScheduler==3.11.0
